refactor(effects): split ai_step.go decode/encode helpers into ai_decode.go + ai_encode.go (S1_M1)
Behaviour-free move — zero logic change. ai_step.go 778 -> 525 lines
(800-line CI gate headroom restored). NOTE: plan ordered the split
BEFORE the patch; reordered because the verbatim patch's context lines
reference the pre-split file — patch-first preserves verbatim
reviewability, split content is identical (reused from the executor's
first pass).

Refs #546
